[features/lane_detection] Updated the basic codes for the lane detection
1. Updated codes for traffic_light.cpp to replace bilateralFilter to Gaussianblur
2. Updated SteeringAngle in driving_control.cpp to return steering_angle and img constantly
3. Updated straightLaneAngle in lane_detection.cpp to return different straight_lane_angle in each traffic_light situation

+++ features/lane_detection/include/lane_detection/camera/traffic_light.h
@@ -9,14 +9,14 @@
 
 class TrafficLight {
 private: // NOLINT
-  int light_color;
+  uint8_t light_color;
   double sigma_color;
   double sigma_space;
 
 public: // NOLINT
   TrafficLight() : light_color(0), sigma_color(10), sigma_space(10) {}
   ~TrafficLight() {}
-  int FindTrafficLight(cv::Mat img);
+  uint8_t FindTrafficLight(cv::Mat img);
 };
 
 #endif /* TRAFFIC_LIGHT_H */
